+++ src/de/omegazirkel/risingworld/tools/settings/AdminSettingsType.java
@@ -3,5 +3,7 @@ package de.omegazirkel.risingworld.tools.settings;
 public enum AdminSettingsType {
     BOOLEAN,
     INTEGER,
+    SELECT,
+    TEXT,
     STRING
 }