+++ src/de/omegazirkel/risingworld/tools/ui/ToolsPlayerPluginSettingsPanel.java
@@ -1,20 +1,34 @@
 package de.omegazirkel.risingworld.tools.ui;
 
+import de.omegazirkel.risingworld.OZTools;
+import de.omegazirkel.risingworld.tools.I18n;
+import de.omegazirkel.risingworld.tools.ToolsPlayerPreferences;
 import net.risingworld.api.objects.Player;
-import net.risingworld.api.ui.UILabel;
 
 public class ToolsPlayerPluginSettingsPanel extends BasePlayerPluginSettingsPanel {
+    private final Player uiPlayer;
+
+    private static I18n t() {
+        return I18n.getInstance(OZTools.name);
+    }
 
     public ToolsPlayerPluginSettingsPanel(Player uiPlayer, String pluginLabel) {
         super(uiPlayer, pluginLabel);
+        this.uiPlayer = uiPlayer;
     }
 
     @Override
     protected void redrawContent() {
         flexWrapper.removeAllChilds();
-        // TODO: implement actual settings content for OZ - Tools plugin
-        UILabel placeholderLabel = new UILabel("OZ - Tools plugin settings will be here.");
-        flexWrapper.addChild(placeholderLabel);
+        OZUIElement labels = defaultSettingsContainer();
+        labels.addChild(defaultSettingsLabel(t().get("TC_TOOLS_SETTING_INVENTORY_LABELS", uiPlayer)));
+        labels.addChild(switchButtons(uiPlayer, ToolsPlayerPreferences.showInventoryShortcutLabels(uiPlayer), event -> {
+            boolean next = !ToolsPlayerPreferences.showInventoryShortcutLabels(uiPlayer);
+            ToolsPlayerPreferences.setShowInventoryShortcutLabels(uiPlayer, next);
+            InventoryOverlayPanel.refreshAllVisible();
+            redrawContent();
+        }));
+        flexWrapper.addChild(labels);
     }
 
 }