+++ src/de/omegazirkel/risingworld/tools/utils/RegionType.java
@@ -1,0 +1,14 @@
+package de.omegazirkel.risingworld.tools.utils;
+
+public enum RegionType {
+    FOREST("Forest"),
+    ARCTIC("Arctic"),
+    DESERT("Desert"),
+    UNKNOWN("Unknown");
+
+    public final String name;
+
+    RegionType(String name) {
+        this.name = name;
+    }
+}